rstsr-blas-traits: move EigenRange to dedicated eigen_range module
Refactor EigenRange enum and its From implementations from syevr.rs
to a new eigen_range.rs module for better code organization and
reuse across multiple LAPACK eigenvalue drivers.
